requirements.txt
streamlit
pandas
openpyxl
requests
python-dotenv
folium
scikit-learn
celery
redis
uvicorn
python-multipart
email_validator
fastapi
psycopg2-binary
supabase
ortools
watchdog
ipykernel
jupyter_core
debugpy
shapely
matplotlib